Cleanup, removal of unnecessary functions, addition of referential concentration
Removed unnecessary reaction rate functions regarding position since they are too specific and this functionality can be introduced via math expressions now.
Introduced referential volume solute concentration which neglects the change of concentration due to strain. This is useful for kinematic growth when we assume that concentrations/reactions only change with the elastic but not growth deformation.